github-app/requirements.txt
# Drift Bot — GitHub App server
# Minimal dependencies for the webhook handler

fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
httpx>=0.27,<1
PyJWT[crypto]>=2.8,<3
cryptography>=42,<45
drift-analyzer
